+++ example/src/main/java/andrewoid/neuralnetwork/DigitUI/DigitFrame.java
@@ -1,4 +1,6 @@
-package andrewoid.neutralnetwork;
+package andrewoid.neuralnetwork.DigitUI;
+
+import andrewoid.neutralnetwork.Network;
 
 import javax.swing.*;
 import java.awt.*;
